backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-RESTful
razorpay
python-dotenv
reportlab